3.2.1. 3.2.2 Wykoananie i dodanie pozostałych testów jednostkowych. Dodatkowo wykonanie punktu 3.1.12
