--- src/main/java/edu/kis/vh/nursery/list/Node.java
@@ -1,35 +1,0 @@
-package edu.kis.vh.nursery.list;
-
-class Node {
-
-    private int value;
-    private Node prev, next;
-
-    public Node(int i) {
-        setValue(i);
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    public void setValue(int value) {
-        this.value = value;
-    }
-
-    public Node getPrev() {
-        return prev;
-    }
-
-    public void setPrev(Node prev) {
-        this.prev = prev;
-    }
-
-    public Node getNext() {
-        return next;
-    }
-
-    public void setNext(Node next) {
-        this.next = next;
-    }
-}
